+++ martypy/ClientMV2.py
@@ -64,6 +64,7 @@ class ClientMV2(ClientGeneric):
         self.publishedMsgCallback = None
         self.reportMsgCallback = None
         self._initComplete = False
+        self.maxWaitForConnReadySecs = 10
         self._minSysVersForSubscribeAPI = "1.0.0"
         self._interfaceMethod = method
         self._numHwStatusRetries = 10
@@ -115,6 +116,15 @@ class ClientMV2(ClientGeneric):
         self._getRICVersion()
         self._updateHwElemsInfo()
         self._initComplete = True
+        # Wait for connection to be ready
+        waitStartTime = time.time()
+        while not self.is_conn_ready():
+            if time.time() - waitStartTime > self.maxWaitForConnReadySecs:
+                raise MartyConnectException("Connection to Marty not ready")
+            time.sleep(0.1)
+        # A flag is set on the first subscribed message - wait a little longer
+        # to ensure that one of each subscribed message type has been received
+        time.sleep(0.5)
 
     def close(self):
         if self.isClosing:
@@ -289,10 +299,101 @@ class ClientMV2(ClientGeneric):
 
     def get_distance_sensor(self) -> int:
         for attached_add_on in self.get_add_ons_status().values():
-            if type(attached_add_on) == dict and attached_add_on['whoAmITypeCode'] == '00000083':
+            if type(attached_add_on) == dict and attached_add_on['whoAmI'] == 'VCNL4200':
                 distance_bytes = attached_add_on['data'][1:3]
                 distance = int.from_bytes(distance_bytes, 'big')
                 return distance
+        return 0
+
+    def _index_data_color_ir(self, attached_add_on: str) -> Tuple[Tuple[int, int, int], str]:
+        '''
+        Parses data of passed-in color or IR sensor :two:
+        Args:
+            attached_add_on: Name of IR or color sensor add on
+        Returns:
+            A tuple with three integers of data (detection byte, obstacle value, ground value)
+            and a string of whether the sensor is a default 'left' or 'right' add on
+        '''
+        if attached_add_on['whoAmI'] == 'IRFoot':
+            detection_flags = attached_add_on['data'][1]
+            obstacle_data_raw = int.from_bytes(attached_add_on['data'][2:4], 'big')
+            ground_data_raw = int.from_bytes(attached_add_on['data'][4:6], 'big')
+            return (detection_flags, obstacle_data_raw, ground_data_raw), 'right'
+        if attached_add_on['whoAmI'] == 'coloursensor':
+            detection_flags = attached_add_on['data'][6]
+            obstacle_data_raw = int.from_bytes(attached_add_on['data'][8:10], 'big')
+            ground_data_raw = attached_add_on['data'][2]
+            return (detection_flags, obstacle_data_raw, ground_data_raw), 'left'
+
+    def _get_obstacle_and_ground_raw_data(self, add_on_or_side: str) -> list:
+        '''
+        Finds the wanted color or IR sensor and gets the parsed obstacle and ground data :two:
+        Args:
+            add_on_or_side: Takes in the name of a color sensor, name of an IR
+            sensor, `'left'` for the add on connected to the left foot,
+            or `'right'` for the add on connected to the right foot
+        Returns:
+            A tuple of the detection byte, obstacle reading, and ground reading of the add on
+        '''
+        sensor_whoamis = {'IRFoot', 'coloursensor'}
+        sensor_data = {'left': [], 'right': []}
+        sensor_possible_names = {
+            'left': ['LeftColorSensor', 'LeftIRFoot'],
+            'right': ['RightIRFoot', 'RightColorSensor']
+        }
+        addon_names = sensor_possible_names.get(add_on_or_side, [])
+
+        # There may be a situation just after connecting to RIC where publication messages from the addon
+        # have not yet been received so we need to wait for them to be received before we can parse them
+        for retryLoop in range(10):
+            addon_statuses = self.get_add_ons_status().values()
+            if len(addon_statuses) != 0:
+                break
+            time.sleep(0.1)
+
+        # Get the add-on values
+        for attached_add_on in addon_statuses:
+            if type(attached_add_on) == dict and attached_add_on.get('whoAmI','') in sensor_whoamis:
+                obstacle_and_ground_data, side = self._index_data_color_ir(attached_add_on)
+                if attached_add_on['name'] in addon_names:
+                    return obstacle_and_ground_data
+                else:
+                    sensor_data[side].append(obstacle_and_ground_data)
+        if len(sensor_data[add_on_or_side.lower()]) == 1:
+            return sensor_data[add_on_or_side.lower()][0]
+        elif add_on_or_side.lower() == 'left' or add_on_or_side.lower() == 'right':
+            raise MartyCommandException(
+                f"Marty could not find a {add_on_or_side} sensor. "
+                "Please make sure your add ons are plugged in and named correctly"
+            )
+        else:
+            raise MartyCommandException(
+                f"The add on '{add_on_or_side}' is not a valid add on for obstacle and ground sensing."
+                "Please make sure to pass in the name of an IR sensor or Color sensor."
+            )
+
+    def foot_on_ground(self, add_on_or_side: str) -> bool:
+        raw_data = self._get_obstacle_and_ground_raw_data(add_on_or_side)
+        if len(raw_data) > 0:
+            return (raw_data[0] & 0b10) == 0b00
+        return False
+
+    def foot_obstacle_sensed(self, add_on_or_side: str) -> bool:
+        raw_data = self._get_obstacle_and_ground_raw_data(add_on_or_side)
+        if len(raw_data) > 0:
+            return (raw_data[0] & 0b1) == 0b1
+        return False
+
+    def get_obstacle_sensor_reading(self, add_on_or_side: str) -> int:
+        raw_data = self._get_obstacle_and_ground_raw_data(add_on_or_side)
+        if len(raw_data) > 1:
+            return raw_data[1]
+        return 0
+
+    def get_ground_sensor_reading(self, add_on_or_side: str) -> int:
+        raw_data = self._get_obstacle_and_ground_raw_data(add_on_or_side)
+        if len(raw_data) > 2:
+            return raw_data[2]
         return 0
 
     def get_accelerometer(self, axis: Optional[str] = None, axisCode: int = 0) -> float:
@@ -413,10 +514,10 @@ class ClientMV2(ClientGeneric):
         return self._initComplete and (self.lastRICSerialMsgTime is not None)
 
     def _is_valid_disco_addon(self, add_on: str) -> bool:
-        disco_type_codes = {"00000087","00000088","00000089"}
+        disco_whoamis = {"LEDfoot", "LEDarm", "LEDeye"}
         for attached_add_on in self.get_add_ons_status().values():
             if type(attached_add_on) == dict and attached_add_on['name'] == add_on:
-                if attached_add_on['whoAmITypeCode'] in disco_type_codes:
+                if attached_add_on['whoAmI'] in disco_whoamis:
                     return True
                 else:
                     raise MartyCommandException(f"The add on name: '{add_on}' is not a valid disco add on. "
@@ -490,7 +591,7 @@ class ClientMV2(ClientGeneric):
             response = self.add_on_query(add_on, command, 0)
             return response.get("rslt", "") == "ok"
 
-    def disco_group_operation(self, disco_operation: Callable, whoami_type_codes: set, operation_kwargs: dict) -> bool:
+    def disco_group_operation(self, disco_operation: Callable, whoamis: set, operation_kwargs: dict) -> bool:
         '''
         Calls disco operations in groups for multiple add ons :two:
         Args:
@@ -502,7 +603,7 @@ class ClientMV2(ClientGeneric):
         '''
         result = True
         for attached_add_on in self.get_add_ons_status().values():
-            if type(attached_add_on) == dict and attached_add_on['whoAmITypeCode'] in whoami_type_codes:
+            if type(attached_add_on) == dict and attached_add_on['whoAmI'] in whoamis:
                 addon_name = attached_add_on['name']
                 result = result and disco_operation(add_on=addon_name, **operation_kwargs)
         return result